# go/model_task_id.go
/*
 * Тестовая служба работы с задачами
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// Идентификатор задачи (UUID)
type TaskId struct {
}