# go/model_task_status.go
/*
 * Тестовая служба работы с задачами
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

// # Состояние задачи 
type TaskStatus struct {
	// ## статус задачи - `created` Задача создана, но ещё не запущена - `running` Задача выполняется - `finished` Задача выполнена 
	Status string `json:"status,omitempty"`
	// ## Дата и время изменения статуса 
	Timestamp time.Time `json:"timestamp,omitempty"`
}